+++ railties/lib/rails/generators/rails/authentication/templates/app/mailers/password_mailer.rb.tt
@@ -1,4 +1,4 @@
-class PasswordsMailer < ApplicationMailer
+class PasswordMailer < ApplicationMailer
   def reset(user)
     @user = user
     mail subject: "Reset your password", to: user.email_address
