+++ railties/lib/rails/generators/rails/authentication/templates/test/mailers/previews/password_mailer_preview.rb.tt
@@ -1,0 +1,7 @@
+# Preview all emails at http://localhost:3000/rails/mailers/password_mailer
+class PasswordMailerPreview < ActionMailer::Preview
+  # Preview this email at http://localhost:3000/rails/mailers/password_mailer/reset
+  def reset
+    PasswordsMailer.reset(User.take)
+  end
+end
