--- railties/lib/rails/generators/rails/authentication/templates/test/mailers/previews/passwords_mailer_preview.rb.tt
@@ -1,7 +1,0 @@
-# Preview all emails at http://localhost:3000/rails/mailers/passwords_mailer
-class PasswordsMailerPreview < ActionMailer::Preview
-  # Preview this email at http://localhost:3000/rails/mailers/passwords_mailer/reset
-  def reset
-    PasswordsMailer.reset(User.take)
-  end
-end
